% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{agepregartcov}
\alias{agepregartcov}
\title{Age-specific ART coverage among pregnant women}
\usage{
agepregartcov(
  mod,
  fp,
  aidx = 3:9 * 5 - fp$ss$AGE_START + 1L,
  yidx = 1:fp$ss$PROJ_YEARS,
  agspan = 5,
  expand = FALSE
)
}
\arguments{
\item{expand}{whether to expand aidx, yidx, sidx, and agspan}
}
\description{
Age-specific ART coverage among pregnant women
}
