% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{ageprev}
\alias{ageprev}
\title{Prevalence by arbitrary age groups}
\usage{
ageprev(
  mod,
  aidx = NULL,
  sidx = NULL,
  yidx = NULL,
  agspan = 5,
  expand = FALSE,
  VERSION = "C"
)
}
\arguments{
\item{sidx}{sex (1 = Male, 2 = Female, 0 = Both)
Notes: Assumes that AGE_START is 15 and single year of age.}
}
\description{
Prevalence by arbitrary age groups
}
