% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infections.R
\name{calc_infections_eppspectrum}
\alias{calc_infections_eppspectrum}
\title{Annualized number of new infections}
\usage{
calc_infections_eppspectrum(fp, pop, hivpop, artpop, i, ii, r_ts)
}
\description{
Annualized number of new infections
}
