% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{incid}
\alias{incid}
\title{Incidence rate among adults age 15-49 years}
\usage{
incid(mod, ...)
}
\arguments{
\item{mod}{model output}
}
\description{
Incidence rate among adults age 15-49 years
}
\details{
This returns incidence rate calculated as the number of infections during
the projection period divded by the number susceptible at the mid-point of
the projection period. This is the default incidence calculation for
Spectrum version >=6.2 For Spectrum versions <=6.19 incidence was calculated
as number of infections divided by susceptible population at the start of the
projection year.
}
