% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ancsite.R
\name{ll_ancsite_conditional}
\alias{ll_ancsite_conditional}
\title{Pointwise likelihood for site-level ANC observations given site-level effects}
\usage{
ll_ancsite_conditional(mod, fp, newdata, b_site)
}
\description{
Pointwise likelihood for site-level ANC observations given site-level effects
}
