% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll_hhsage}
\alias{ll_hhsage}
\title{Log likelihood for age-specific household survey prevalence}
\usage{
ll_hhsage(mod, fp, dat, pointwise = FALSE)
}
\description{
Log likelihood for age-specific household survey prevalence
}
