% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll_hhsartcov}
\alias{ll_hhsartcov}
\title{Log likelihood for age-specific household survey prevalence}
\usage{
ll_hhsartcov(mod, dat, pointwise = FALSE)
}
\description{
Log likelihood for age-specific household survey prevalence
}
