% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll_hhsincid}
\alias{ll_hhsincid}
\title{Log-likelhood for direct incidence estimate from household survey}
\usage{
ll_hhsincid(mod, fp, hhsincid.dat)
}
\arguments{
\item{mod}{model output, object of class `spec`.}

\item{hhsincid.dat}{prepared houshold survey incidence estimates (see perp}
}
\description{
Calculate log-likelihood for nationally representative incidence
estimates from a household survey. Currently implements likelihood
for a log-transformed direct incidence estimate and standard error.
Needs to be updated to handle incidence assay outputs.
}
