% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppasm-outputs.R
\name{mod_dimnames}
\alias{mod_dimnames}
\title{Add dimnames to spec model output}
\usage{
mod_dimnames(mod, ss)
}
\arguments{
\item{mod}{EPP-ASM model output of class `spec`}

\item{ss}{Model state space inputs}
}
\value{
EPP-ASM model output with labelled dimensions
}
\description{
`mod_dimnames` assigns dimnames to spec model outputs.
}
