% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{natagemx.spec}
\alias{natagemx.spec}
\title{Non-HIV age-specific mortality}
\usage{
\method{natagemx}{spec}(mod)
}
\arguments{
\item{mod}{output of simmod of class \code{\link{spec}}.}
}
\value{
3-dimensional array of mortality by age, sex, and year.
}
\description{
Calculate all-cause mortality rate by single year of age and sex from a
\code{spec} object.
}
\details{
Mortality in year Y is calculated as the number of non-HIV deaths occurring
from the mid-year of year Y-1 to mid-year Y, divided by the population size
at the mid-year of year Y-1.
!!! NOTE: This might be different from the calculation in Spectrum. Should
confirm this with John Stover.
}
