% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ancmod.R
\name{prepare_anc_model}
\alias{prepare_anc_model}
\title{Prepare ANC model object}
\usage{
prepare_anc_model(
  fp,
  eppd,
  fit_vinfl = fp$fit_vinfl,
  fit_ancrtcens_vinfl = fp$fit_ancrtcens_vinfl,
  fit_ancrtsite_beta = fp$fit_ancrtsite_beta
)
}
\description{
Prepare ANC model object
}
\details{
Defaults are to:
* Fit site-level ANC bias parameter if any ANC data exist
* Fit ANC-RT vs. ANC-SS site-level bias if ANC-RT site data exist
* Fit fertility rate ratio adjustment if ANC-RT census data exist
* Fit vinfl and ancrtcens_vinfl if ANC site or ANC data data exist, respectively
}
