% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directincid.R
\name{prepare_directincid}
\alias{prepare_directincid}
\title{Prepare direct incidence input EPP simulation}
\usage{
prepare_directincid(pjnz, method = "directincid_hts")
}
\arguments{
\item{method}{Method for intercalating new infections. Either "directincid_ann"
for annual new infections or "directincid_hts" for new infections each
HIV time step.}
}
\description{
Prepare direct incidence input EPP simulation
}
