% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_hhsageprev_likdat}
\alias{prepare_hhsageprev_likdat}
\title{Prepare age-specific HH survey prevalence likelihood data}
\usage{
prepare_hhsageprev_likdat(hhsage, fp)
}
\description{
Prepare age-specific HH survey prevalence likelihood data
}
