% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_hhsartcov_likdat}
\alias{prepare_hhsartcov_likdat}
\title{## Household survey ART coverage likelihood}
\usage{
prepare_hhsartcov_likdat(hhsartcov, fp)
}
\description{
## Household survey ART coverage likelihood
}
