% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_hhsincid_likdat}
\alias{prepare_hhsincid_likdat}
\title{Prepare household survey incidence likelihood data}
\usage{
prepare_hhsincid_likdat(hhsincid, fp)
}
\description{
Prepare household survey incidence likelihood data
}
