% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt-models.R
\name{prepare_rhybrid}
\alias{prepare_rhybrid}
\title{Setup the r-hybrid model}
\usage{
prepare_rhybrid(
  fp,
  tsEpidemicStart = fp$ss$time_epi_start + 0.5,
  rw_start = fp$rw_start,
  rw_trans = fp$rw_trans,
  rw_dk = fp$rw_dk
)
}
\arguments{
\item{fp}{model parameters object}

\item{tsEpidemicStart}{time step at which epidemic is seeded}

\item{rw_start}{time when random walk starts}

\item{rw_trans}{number of years to transition from logistic differences to RW differences. If NULL, defaults to 5 steps}
}
\description{
Setup the r-hybrid model
}
