% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-model.R
\name{prepare_spec_fit}
\alias{prepare_spec_fit}
\title{Prepare fit by EPP regions}
\usage{
prepare_spec_fit(
  pjnz,
  proj.end = 2016.5,
  popadjust = NULL,
  popupdate = TRUE,
  use_ep5 = FALSE
)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}

\item{proj.end}{end year for projection.}

\item{popupdate}{logical should target population be updated to match age-specific population size from DP file and %Urban from EPP XML.}
}
\description{
Prepare fit by EPP regions
}
