% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-files.R
\name{read_csavr_data}
\alias{read_csavr_data}
\title{Read CSAVR input data}
\usage{
read_csavr_data(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\description{
Read CSAVR input data
}
