% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-files.R
\name{read_dp}
\alias{read_dp}
\title{Read Spectrum internal files file from PJNZ}
\usage{
read_dp(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\description{
Read Spectrum internal files file from PJNZ
}
