% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-files.R
\name{read_incid_input}
\alias{read_incid_input}
\title{Read annual incidence input}
\usage{
read_incid_input(pjnz)
}
\arguments{
\item{pjnz}{file path to Spectrum PJNZ file.}
}
\description{
Read annual incidence input
}
