% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-spectrum-pop1.R
\name{read_pop1}
\alias{read_pop1}
\title{Read Spectrum _pop1.xlsx export file}
\usage{
read_pop1(pop1file, country, years = 2000:2021)
}
\arguments{
\item{pop1file}{path to *_pop1.xlsx export file}

\item{country}{string giving the identifying country / region
name to be added as a column in the output file}

\item{years}{integer vector of years to extract}
}
\value{
A data.frame consisting of pop1 file in long format
}
\description{
Spectrum has a debug mode option to output a file recording
the entire state space when a model simulation is computed.
This function parse the *_pop1.xlsx file into long-format
data.frame.
}
\details{
TODO: tidy up the dependencies in this function: reshape, pryr, readxl.

TODO: Insert instructions for how to enter Spectrum debug mode.

`years` must be in the file as sheets, but this is not currently checked.
It will fail hard. Inserting a check for this would make it fail softer.
}
