% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ancmod.R
\name{sample_ancsite_pred}
\alias{sample_ancsite_pred}
\title{Sample posterior predictions for site-level ANC observations}
\usage{
sample_ancsite_pred(mod, fp, newdata, b_site)
}
\description{
Sample posterior predictions for site-level ANC observations
}
