% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ancmod.R
\name{sample_b_site}
\alias{sample_b_site}
\title{Sample from posterior ditribution for ANC site level random effects}
\usage{
sample_b_site(mod, fp, dat, resid = TRUE)
}
\description{
Sample from posterior ditribution for ANC site level random effects
}
