% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt-models.R
\name{sample_invgamma_post}
\alias{sample_invgamma_post}
\title{Sample from conditional posterior distribution for variance parameter}
\usage{
sample_invgamma_post(x, prior_shape, prior_rate)
}
\description{
Sample from conditional posterior distribution for variance parameter
}
