% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppasm.R
\name{spec_add_dimnames}
\alias{spec_add_dimnames}
\title{Add dimnames to EPP-ASM model output}
\usage{
spec_add_dimnames(mod, fp)
}
\arguments{
\item{mod}{output from `simmod()`}

\item{fp}{fixed parameters input to `simmod()`}
}
\value{
Input `mod` object with dimnames applied to arrays.
}
\description{
Add dimnames to EPP-ASM model output
}
