% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_aggr.R
\name{tidy_aggr}
\alias{tidy_aggr}
\title{Aggregate a list of fits}
\usage{
tidy_aggr(fitlist, modlab = NA, geolab = NA)
}
\description{
Aggregate a list of fits
}
\details{
This makes some assumptions:

* The number of resamples is the same for all fits in list
* The state space dimension, including number of projection years, is the same
}
