% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{prepare_ancsite_likdat}
\alias{prepare_ancsite_likdat}
\title{Prepare site-level ANC prevalence data for EPP random-effects likelihood}
\usage{
prepare_ancsite_likdat(ancsitedat, fp)
}
\arguments{
\item{ancsitedat}{data.frame of site-level ANC data}

\item{fp}{fixed parameter input list, including state space}
}
\description{
Prepare site-level ANC prevalence data for EPP random-effects likelihood
}
